Move account to feature tier
account owns the /account/* route, so it's a feature, not shared. Flip
its tier in module-boundaries.mjs and resolve the two violations the
promotion exposes — both the shared notifications module reaching into
account for notification *concepts* (not account-page concepts):

- NotificationListItem (+ union members, NotificationType): moved the type
  file account/notifications/types.ts -> notifications/types.ts.
- formatCreatorMessage: moved into the shared notifications/utils.ts and
  deleted the now-empty account NotificationCard/utils.ts.

All importers (account consumes the concept downward) repointed. No
behavior change.

DEV-2134
