+++ frontend/src/metabase/account/notifications/components/NotificationCard/DashboardNotificationCard.tsx
@@ -2,10 +2,12 @@ import type { JSX } from "react";
 import { useCallback } from "react";
 import { t } from "ttag";
 
-import { formatCreatorMessage } from "metabase/account/notifications/components/NotificationCard/utils";
-import type { DashboardSubscriptionListItem } from "metabase/account/notifications/types";
 import { Link } from "metabase/common/components/Link";
-import { formatTitle } from "metabase/notifications/utils";
+import type { DashboardSubscriptionListItem } from "metabase/notifications/types";
+import {
+  formatCreatorMessage,
+  formatTitle,
+} from "metabase/notifications/utils";
 import { canArchiveLegacyAlert, formatChannel } from "metabase/pulse";
 import { Box, Flex, Icon, Text } from "metabase/ui";
 import * as Urls from "metabase/urls";
