+++ frontend/src/metabase/account/notifications/components/NotificationCard/NotificationCard.tsx
@@ -1,11 +1,11 @@
 import { type JSX, useCallback } from "react";
 import { t } from "ttag";
 
-import { formatCreatorMessage } from "metabase/account/notifications/components/NotificationCard/utils";
-import type { QuestionNotificationListItem } from "metabase/account/notifications/types";
 import { Link } from "metabase/common/components/Link/Link";
+import type { QuestionNotificationListItem } from "metabase/notifications/types";
 import {
   canArchive,
+  formatCreatorMessage,
   formatNotificationSchedule,
   formatTitle,
   getNotificationEnabledChannelsMap,
