+++ frontend/src/metabase/notifications/modals/QuestionAlertListModal/AlertListItem.tsx
@@ -3,8 +3,8 @@ import type { MouseEvent } from "react";
 import { msgid, ngettext, t } from "ttag";
 import _ from "underscore";
 
-import { formatCreatorMessage } from "metabase/account/notifications/components/NotificationCard/utils";
 import {
+  formatCreatorMessage,
   formatNotificationSchedule,
   getNotificationHandlersGroupedByTypes,
 } from "metabase/notifications/utils";
