+++ frontend/src/metabase/notifications/utils.ts
@@ -1,8 +1,8 @@
 import { c, msgid, ngettext, t } from "ttag";
 import _ from "underscore";
 
-import type { NotificationListItem } from "metabase/account/notifications/types";
 import { cronToScheduleSettings } from "metabase/common/components/Schedule/cron";
+import type { NotificationListItem } from "metabase/notifications/types";
 import { getScheduleExplanation } from "metabase/utils/cron";
 import { getEmailDomain, isEmail } from "metabase/utils/email";
 import { formatTimeWithUnit } from "metabase/utils/formatting/time";
@@ -151,6 +151,29 @@ export const formatTitle = ({ item, type }: NotificationListItem) => {
     case "question-notification":
       return item.payload?.card?.name || t`Alert`;
   }
+};
+
+export const formatCreatorMessage = (
+  item: NotificationListItem["item"],
+  userId: UserId,
+) => {
+  let creatorString = "";
+  const options = MetabaseSettings.formattingOptions();
+
+  if (userId === item.creator?.id) {
+    creatorString += t`Created by you`;
+  } else if (item.creator?.common_name) {
+    creatorString += t`Created by ${item.creator.common_name}`;
+  } else {
+    creatorString += t`Created`;
+  }
+
+  if (item.created_at) {
+    const createdAt = formatDateTimeWithUnit(item.created_at, "day", options);
+    creatorString += t` on ${createdAt}`;
+  }
+
+  return creatorString;
 };
 
 const getRecipientIdentity = (recipient: NotificationRecipient) => {
